New and moved files, and associated include changes.
