+++ src/osvr/Client/AnalogRemoteFactory.h
@@ -30,7 +30,7 @@
 #include <osvr/Common/InterfaceList.h>
 #include <osvr/Common/OriginalSource.h>
 #include <osvr/Util/SharedPtr.h>
-#include "RemoteHandler.h"
+#include <osvr/Client/RemoteHandler.h>
 
 #include <osvr/Common/ClientContext.h>
