+++ src/osvr/Client/ButtonRemoteFactory.cpp
@@ -30,7 +30,7 @@
 #include <osvr/Util/ChannelCountC.h>
 #include <osvr/Util/UniquePtr.h>
 #include <osvr/Common/OriginalSource.h>
-#include "InterfaceTree.h"
+#include <osvr/Client/InterfaceTree.h>
 #include <osvr/Util/ValueOrRange.h>
 #include <osvr/Util/Verbosity.h>
