+++ src/osvr/Client/CMakeLists.txt
@@ -1,9 +1,13 @@
 osvr_setup_lib_vars(Client)
 
 set(API
+    "${HEADER_LOCATION}/ClientObjectsAndCallbacks.h"
     "${HEADER_LOCATION}/CreateContext.h"
     "${HEADER_LOCATION}/DisplayConfig.h"
+    "${HEADER_LOCATION}/HandlerContainer.h"
     "${HEADER_LOCATION}/InternalInterfaceOwner.h"
+    "${HEADER_LOCATION}/InterfaceTree.h"
+    "${HEADER_LOCATION}/RemoteHandler.h"
     "${HEADER_LOCATION}/Viewer.h"
     "${HEADER_LOCATION}/ViewerEye.h"
     "${HEADER_LOCATION}/ViewerEyeSurface.h")
@@ -13,6 +17,7 @@ set(SOURCE
     AnalogRemoteFactory.h
     ButtonRemoteFactory.cpp
     ButtonRemoteFactory.h
+    ClientObjectsAndCallbacks.cpp
     CreateContext.cpp
     DirectionRemoteFactory.cpp
     DirectionRemoteFactory.h
@@ -21,17 +26,14 @@ set(SOURCE
     DisplayDescriptorSchema1.h
     EyeTrackerRemoteFactory.cpp
     EyeTrackerRemoteFactory.h
-    HandlerContainer.h
     ImagingRemoteFactory.cpp
     ImagingRemoteFactory.h
     InterfaceTree.cpp
-    InterfaceTree.h
     Location2DRemoteFactory.cpp
     Location2DRemoteFactory.h
     PureClientContext.cpp
     PureClientContext.h
     RemoteHandler.cpp
-    RemoteHandler.h
     RemoteHandlerFactory.cpp
     RemoteHandlerFactory.h
     RouterPredicates.h
