+++ src/osvr/Client/ClientObjectsAndCallbacks.cpp
@@ -14,7 +14,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-// 	http://www.apache.org/licenses/LICENSE-2.0
+//        http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -23,16 +23,10 @@
 // limitations under the License.
 
 // Internal Includes
-#include <osvr/Client/RemoteHandler.h>
+#include <osvr/Client/ClientObjectsAndCallbacks.h>
 
 // Library/third-party includes
 // - none
 
 // Standard includes
 // - none
-
-namespace osvr {
-namespace client {
-    RemoteHandler::~RemoteHandler() {}
-} // namespace client
-} // namespace osvr
