+++ src/osvr/Client/Location2DRemoteFactory.h
@@ -31,7 +31,7 @@
 #include <osvr/Common/InterfaceList.h>
 #include <osvr/Common/OriginalSource.h>
 #include <osvr/Util/SharedPtr.h>
-#include "RemoteHandler.h"
+#include <osvr/Client/RemoteHandler.h>
 #include <osvr/Common/ClientContext.h>
 
 // Library/third-party includes
